Move a bunch of models into @weco/catalogue
