+++ catalogue/webapp/components/ArchiveTree/ArchiveTree.tsx
@@ -15,7 +15,7 @@ import Space from '@weco/common/views/components/styled/Space';
 import WorkTitle from '../WorkTitle/WorkTitle';
 import Icon from '@weco/common/views/components/Icon/Icon';
 import { getArchiveAncestorArray } from '../../utils/works';
-import { RelatedWork, Work } from '@weco/common/model/catalogue';
+import { RelatedWork, Work } from '@weco/catalogue/services/catalogue/types';
 import Modal from '@weco/common/views/components/Modal/Modal';
 import ButtonSolid from '@weco/common/views/components/ButtonSolid/ButtonSolid';
 import IsArchiveContext from '../IsArchiveContext/IsArchiveContext';
