+++ catalogue/webapp/components/ExpandedImage/ExpandedImage.tsx
@@ -7,7 +7,7 @@ import {
   getProductionDates,
 } from '@weco/catalogue/utils/works';
 import { getCatalogueLicenseData } from '@weco/common/utils/licenses';
-import { Image as ImageType, Work } from '@weco/common/model/catalogue';
+import { Image as ImageType, Work } from '@weco/catalogue/services/catalogue/types';
 import { getWorkClientSide } from '@weco/catalogue/services/catalogue/works';
 import { expandedViewImageButton } from '@weco/common/text/aria-labels';
 import { fetchIIIFPresentationManifest } from '@weco/catalogue/services/iiif/fetch/manifest';
