+++ catalogue/webapp/components/WorkTabbedNav/WorkTabbedNav.tsx
@@ -1,4 +1,4 @@
-import { Work } from '@weco/common/model/catalogue';
+import { Work } from '@weco/catalogue/services/catalogue/types';
 import { FunctionComponent } from 'react';
 import SubNavigation from '@weco/common/views/components/SubNavigation/SubNavigation';
 import { listView, eye } from '@weco/common/icons';
