+++ catalogue/webapp/pages/api/works/items/[workId].tsx
@@ -1,5 +1,5 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import { ItemsList, CatalogueApiError } from '@weco/common/model/catalogue';
+import { ItemsList, CatalogueApiError } from '@weco/catalogue/services/catalogue/types';
 import {
   catalogueApiError,
   rootUris,
