+++ catalogue/webapp/pages/works/[workId]/images.tsx
@@ -1,7 +1,7 @@
 import { FunctionComponent } from 'react';
 import { GetServerSideProps } from 'next';
 import { appError, AppErrorProps } from '@weco/common/services/app';
-import { Work, Image } from '@weco/common/model/catalogue';
+import { Work, Image } from '@weco/catalogue/services/catalogue/types';
 import { toLink as imageLink } from '@weco/common/views/components/ImageLink/ImageLink';
 import CataloguePageLayout from '@weco/catalogue/components/CataloguePageLayout/CataloguePageLayout';
 import Layout12 from '@weco/common/views/components/Layout12/Layout12';
