+++ catalogue/webapp/services/catalogue/index.ts
@@ -4,7 +4,7 @@ import {
   CatalogueApiError,
   CatalogueResultsList,
   ResultType,
-} from '@weco/common/model/catalogue';
+} from '@weco/catalogue/services/catalogue/types';
 import { Toggles } from '@weco/toggles';
 import { propsToQuery } from '@weco/common/utils/routes';
 import { isString } from '@weco/common/utils/array';
