+++ catalogue/webapp/services/catalogue/types.ts
@@ -1,0 +1,278 @@
+import {
+  Location,
+  DigitalLocation,
+  Identifier,
+  PhysicalLocation,
+  Agent,
+  Contributor,
+} from '@weco/common/model/catalogue';
+
+export type Work = {
+  type: 'Work' | 'Collection' | 'Section' | 'Series';
+  id: string;
+  title: string;
+  alternativeTitles: string[];
+  referenceNumber?: string;
+  description?: string;
+  physicalDescription: string;
+  workType: WorkType;
+  lettering?: string;
+  createdDate?: Period;
+  contributors: Contributor[];
+  identifiers: Identifier[];
+  subjects: Subject[];
+  genres: Genre[];
+  thumbnail?: DigitalLocation;
+  items?: Item<Location>[];
+  production: Production[];
+  currentFrequency?: string;
+  formerFrequency: string[];
+  designation: string[];
+  languages: Language[];
+  edition?: string;
+  notes: Note[];
+  duration?: number;
+  images?: ImageInclude[];
+  parts: RelatedWork[];
+  partOf: RelatedWork[];
+  precededBy: RelatedWork[];
+  succeededBy: RelatedWork[];
+  totalParts?: number;
+  totalDescendentParts?: number;
+  availableOnline?: boolean;
+  availabilities?: Availability[];
+  holdings: Holding[];
+};
+
+export type ItemsList = {
+  type: 'ItemsList';
+  totalResults: number;
+  results: Item<Location>[];
+};
+
+export type Holding = {
+  note?: string;
+  enumeration: string[];
+  location?: Location;
+  type: 'Holdings';
+};
+
+type MinimalRelatedWorkFields =
+  | 'id'
+  | 'title'
+  | 'alternativeTitles'
+  | 'referenceNumber'
+  | 'availableOnline'
+  | 'availabilities'
+  | 'type';
+export type RelatedWork = Partial<Work> & Pick<Work, MinimalRelatedWorkFields>;
+
+type WorkType = {
+  id: string;
+  label: string;
+  type: 'Format';
+};
+
+type Period = {
+  id?: string;
+  identifiers?: Identifier[];
+  label: string;
+  type: 'Period';
+};
+
+type Subject = {
+  id?: string;
+  identifiers?: Identifier[];
+  label: string;
+  concepts: Concept[];
+  type: 'Subject';
+};
+
+type Genre = {
+  label: string;
+  concepts: Concept[];
+  type: 'Genre';
+};
+
+type ConceptType =
+  | 'Subject'
+  | 'Meeting'
+  | 'Organisation'
+  | 'Person'
+  | 'Concept'
+  | 'Period'
+  | 'Place';
+
+export type Concept = {
+  id?: string;
+  identifiers?: Identifier[];
+  label: string;
+  type: ConceptType;
+};
+
+type Availability = {
+  id: string;
+  label: string;
+  type: 'Availability';
+};
+
+export type Item<LocationType> = {
+  id?: string;
+  identifiers?: Identifier[];
+  title?: string;
+  locations: LocationType[];
+  type: 'Item';
+  note?: string;
+};
+
+export type PhysicalItem = Item<PhysicalLocation> & {
+  status?: {
+    id: string;
+    label: string;
+    type: string;
+  };
+};
+
+type Date = {
+  label: string;
+  type: 'Period';
+};
+
+type Place = {
+  id?: string;
+  identifiers?: Identifier[];
+  label: string;
+  type: 'Place';
+};
+
+type Production = {
+  label: string;
+  places: Place[];
+  agents: Agent[];
+  dates: Date[];
+  type: 'ProductionEvent';
+};
+
+type Language = {
+  id?: string;
+  label: string;
+  type: 'Language';
+};
+
+type Note = {
+  contents: string[];
+  noteType: NoteType;
+  type: 'Note';
+};
+
+type NoteType = {
+  id: string;
+  label: string;
+  type: 'NoteType';
+};
+
+type ImageInclude = {
+  id: string;
+  type: 'Image';
+};
+
+// Response objects
+export type CatalogueApiError = {
+  errorType: string;
+  httpStatus: number;
+  label: string;
+  description: string;
+  type: 'Error';
+};
+
+export type CatalogueApiRedirect = {
+  status: number;
+  redirectToId: string;
+  type: 'Redirect';
+};
+
+export type Image = {
+  type: 'Image';
+  id: string;
+  locations: DigitalLocation[];
+  source: {
+    id: string;
+    title: string;
+    contributors?: Contributor[];
+    type: string;
+  };
+  visuallySimilar?: Image[];
+  aspectRatio?: number;
+};
+
+type CatalogueAggregationBucket = {
+  count: number;
+  data: {
+    id: string;
+    label: string;
+    type: string;
+  };
+  type: 'AggregationBucket';
+};
+
+type CatalogueAggregationBucketNoId = {
+  count: number;
+  data: {
+    label: string;
+    type: string;
+  };
+  type: 'AggregationBucket';
+};
+
+type CatalogueAggregation = {
+  buckets: CatalogueAggregationBucket[];
+  type: 'Aggregation';
+};
+
+type CatalogueAggregationNoId = {
+  buckets: CatalogueAggregationBucketNoId[];
+  type: 'Aggregation';
+};
+
+type WorkAggregations = {
+  workType: CatalogueAggregation;
+  availabilities: CatalogueAggregation;
+  languages?: CatalogueAggregation;
+  'genres.label'?: CatalogueAggregationNoId;
+  'subjects.label'?: CatalogueAggregationNoId;
+  'contributors.agent.label'?: CatalogueAggregationNoId;
+  type: 'Aggregations';
+};
+
+type ImageAggregations = {
+  license?: CatalogueAggregation;
+  'source.genres.label'?: CatalogueAggregation;
+  'source.subjects.label'?: CatalogueAggregation;
+  'source.contributors.agent.label'?: CatalogueAggregation;
+  type: 'Aggregations';
+};
+
+type ConceptAggregations = null;
+
+export type ResultType = Work | Image | Concept;
+
+export type CatalogueResultsList<Result extends ResultType> = {
+  type: 'ResultList';
+  totalResults: number;
+  totalPages: number;
+  results: Result[];
+  pageSize: number;
+  prevPage: string | null;
+  nextPage: string | null;
+  aggregations?: Result extends Work
+    ? WorkAggregations
+    : Result extends Image
+    ? ImageAggregations
+    : Result extends Concept
+    ? ConceptAggregations
+    : null;
+
+  // We include the URL used to fetch data from the catalogue API for
+  // debugging purposes.
+  _requestUrl: string;
+};
