+++ catalogue/webapp/utils/json-ld.ts
@@ -1,4 +1,4 @@
-import { Work as WorkType } from '@weco/common/model/catalogue';
+import { Work as WorkType } from '@weco/catalogue/services/catalogue/types';
 import { objToJsonLd } from '@weco/common/utils/json-ld';
 import { JsonLdObj } from '@weco/common/views/components/JsonLd/JsonLd';
